+++ src/app/component/conversation-detail/conversation-emojii/conversation-emojii.component.ts
@@ -1,10 +1,32 @@
-import { Component } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 
 @Component({
   selector: 'chat-conversation-emojii',
   templateUrl: './conversation-emojii.component.html',
   styleUrls: ['./conversation-emojii.component.scss']
 })
-export class ConversationEmojiiComponent {
+export class ConversationEmojiiComponent implements OnInit {
+
+  @Input() var: string;
+  @Output() addEmoji = new EventEmitter();
+  
+  emojiiOptions = {
+    emojiPerLine : 9,
+    totalFrequentLines: 1,
+    showPreview: false,
+    darkMode: false,
+    enableSearch: false,
+    include: [ 'recent', 'people', 'nature', 'activity', 'flags']
+  }
+
+  constructor(){}
+
+  ngOnInit(): void {
+    console.log('varrrrr', this.var)
+  }
+
+  addEmojiFN(event){
+    this.addEmoji.emit(event)
+  }
 
 }